add feature for showing operations
This commit adds functionality for automatically adding all operations
of the tflite to a array test/tflite_ops.json, and prints the operations
from the mcu and saves them in measure_inference_time.py via a lookup table
for the opcodes in test/c_ops.json. The lookup table is necessary as the
operator names in the c code are implemented as enums and the dictionary
is a simple mapping from enum value to the name.